feat: added US03

+++ arqcp/US103/set_sensor_summary_register.h
@@ -1,4 +1,4 @@
 #ifndef SUMMARY_REGISTER
 #define SUMMARY_REGISTER
-void set_sensor_summary_register(int **ptr, int num_sensors, int num_registers, double *res);
+void set_sensor_summary_register(int **ptr, int num_sensors, int num_registers, int *res);
 #endif
